requirements.txt
python-dotenv==1.0.0
line-bot-sdk>=3.5.0
google-generativeai==0.3.0
Pillow==10.1.0
python-multipart==0.0.6
Flask==3.0.0
gunicorn==21.2.0
python-json-logger==2.0.7
